fix: update imports after event refactor
