+++ tests/integration/end_to_end_workflow_test.rs
@@ -12,7 +12,8 @@
 //! 6. **Event-Driven Orchestration** - Event system coordination
 
 use datafold::fold_db_core::infrastructure::message_bus::{
-    MessageBus, FieldValueSetRequest, FieldValueSetResponse
+    MessageBus,
+    request_events::{FieldValueSetRequest, FieldValueSetResponse},
 };
 use datafold::fold_db_core::transform_manager::{TransformManager, TransformUtils};
 use datafold::fold_db_core::managers::atom::AtomManager;
