+++ tests/integration/regression_prevention_test.rs
@@ -12,7 +12,9 @@
 //! 6. **Data Consistency Safeguards** - Ensure data integrity under all conditions
 
 use datafold::fold_db_core::infrastructure::message_bus::{
-    MessageBus, FieldValueSetRequest, FieldValueSetResponse, TransformTriggered, TransformExecuted
+    MessageBus,
+    request_events::{FieldValueSetRequest, FieldValueSetResponse},
+    schema_events::{TransformTriggered, TransformExecuted},
 };
 use datafold::fold_db_core::transform_manager::{TransformManager, TransformUtils};
 use datafold::fold_db_core::managers::atom::AtomManager;
