+++ tests/production_atomref_scenario_test.rs
@@ -6,7 +6,8 @@
 //! - AtomRef UUID: affc56e5-17fb-48ad-b73e-353fe1739d7e (should point to new atom)
 
 use datafold::fold_db_core::infrastructure::message_bus::{
-    MessageBus, FieldValueSetRequest, FieldValueSetResponse
+    MessageBus,
+    request_events::{FieldValueSetRequest, FieldValueSetResponse},
 };
 use datafold::fold_db_core::managers::atom::AtomManager;
 use datafold::db_operations::DbOperations;
