+++ tests/query_layer_atomref_bug_test.rs
@@ -6,7 +6,8 @@
 //! 3. Result: Query finds old/wrong atom UUIDs
 
 use datafold::fold_db_core::infrastructure::message_bus::{
-    MessageBus, FieldValueSetRequest, FieldValueSetResponse
+    MessageBus,
+    request_events::{FieldValueSetRequest, FieldValueSetResponse},
 };
 use datafold::fold_db_core::managers::atom::AtomManager;  
 use datafold::fold_db_core::transform_manager::utils::TransformUtils;
