+++ tests/range_key_query_fix_test.rs
@@ -5,7 +5,8 @@
 
 use datafold::db_operations::DbOperations;
 use datafold::fold_db_core::infrastructure::message_bus::{
-    MessageBus, FieldValueSetRequest, FieldValueSetResponse
+    MessageBus,
+    request_events::{FieldValueSetRequest, FieldValueSetResponse},
 };
 use datafold::fold_db_core::transform_manager::utils::TransformUtils;
 use datafold::fold_db_core::managers::atom::AtomManager;
